Security

+++ src/main/java/by/akimova/CartAPI/service/impl/CartServiceImpl.java
@@ -1,14 +1,19 @@
 package by.akimova.CartAPI.service.impl;
 
 import by.akimova.CartAPI.exception.EntityNotFoundException;
-import by.akimova.CartAPI.exception.ValidationException;
+import by.akimova.CartAPI.exception.NotAccessException;
+import by.akimova.CartAPI.exception.NotValidUsernameException;
 import by.akimova.CartAPI.model.Cart;
 import by.akimova.CartAPI.model.Item;
+import by.akimova.CartAPI.model.User;
 import by.akimova.CartAPI.repository.CartRepository;
 import by.akimova.CartAPI.repository.ItemRepository;
+import by.akimova.CartAPI.repository.UserRepository;
 import by.akimova.CartAPI.service.CartService;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
 import java.util.*;
@@ -19,16 +24,17 @@ import java.util.*;
  * @author anastasiyaakimava
  * @version 1.0
  */
-@AllArgsConstructor
 @Service
 @Slf4j
+@AllArgsConstructor
 public class CartServiceImpl implements CartService {
 
     private final CartRepository cartRepository;
+    private final UserRepository userRepository;
     private final ItemRepository itemRepository;
 
     /**
-     * The method get all carts with all information about it.
+     * The method show all carts with all information about it.
      *
      * @return list of carts.
      */
@@ -41,22 +47,26 @@ public class CartServiceImpl implements CartService {
     }
 
     /**
-     * The method get cart by id with all information about it.
+     * The method show cart with all information about it.
      *
      * @param cartId This is cart's id.
      * @return found cart.
      */
     @Override
-    public Cart getCartById(UUID cartId) throws EntityNotFoundException, ValidationException {
+    public Cart getCartById(UUID cartId) throws NotValidUsernameException, EntityNotFoundException {
+
         if (cartId == null) {
             log.error("IN getCartById - id is null ");
-            throw new ValidationException("cartId is null");
+            throw new NotValidUsernameException("cartId is null");
         }
+
         Cart cart = cartRepository.findCartByCartId(cartId);
+
         if (cart == null) {
             log.error("IN getCartById - no cart found by id: {}", cartId);
             throw new EntityNotFoundException("Cart not found");
         }
+
         log.info("IN getCartById - cart: {} found by id: {}", cart, cartId);
         return cart;
     }
@@ -68,16 +78,20 @@ public class CartServiceImpl implements CartService {
      * @return found cart.
      */
     @Override
-    public Cart getCartByUserId(UUID userId) throws EntityNotFoundException, ValidationException {
+    public Cart getCartByUserId(UUID userId) throws EntityNotFoundException, NotValidUsernameException {
+
         if (userId == null) {
             log.error("IN getCartByUserId - id is null ");
-            throw new ValidationException("cartId is null ");
+            throw new NotValidUsernameException("cartId is null ");
         }
+
         Cart cart = cartRepository.findCartByUserId(userId);
+
         if (cart == null) {
             log.error("IN getCartByUserId - no cart found by userId: {}", userId);
             throw new EntityNotFoundException("Cart doesn't exist ");
         }
+
         log.info("IN getCartByUserId - cart: {} found by userId: {}", cart, userId);
         return cart;
     }
@@ -90,23 +104,46 @@ public class CartServiceImpl implements CartService {
      * @return cart without items which needed to be removed.
      */
     @Override
-    public Cart deleteFromCart(UUID cartId, List<UUID> itemIds) throws EntityNotFoundException, ValidationException {
+    public Cart deleteFromCart(UUID cartId, List<UUID> itemIds) throws EntityNotFoundException, NotValidUsernameException, NotAccessException {
+
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String username;
+
+        if (principal instanceof UserDetails) {
+            username = ((UserDetails) principal).getUsername();
+        } else {
+            username = principal.toString();
+        }
+
         if (cartId == null) {
             log.error("IN deleteFromCart - id is null ");
-            throw new ValidationException("cartId is null");
+            throw new NotValidUsernameException("cartId is null");
         }
+
         Cart dbCart = cartRepository.findCartByCartId(cartId);
+
         if (dbCart == null) {
             log.error("IN deleteFromCart - no cart found by cartId: {}", cartId);
             throw new EntityNotFoundException("Cart doesn't exist ");
         }
-        Collection<Item> items = dbCart.getItems();
-        Collection<Item> toDelete = new ArrayList<>();
-        for (Item item : items) {
-            if (itemIds.contains(item.getItemId()))
-                toDelete.add(item);
+
+        User user = userRepository.findUserByUserId(dbCart.getUserId());
+
+        if (Objects.equals(username, user.getMail())) {
+            Collection<Item> items = dbCart.getItems();
+            Collection<Item> toDelete = new ArrayList<>();
+
+            for (Item item : items) {
+                if (itemIds.contains(item.getItemId()))
+                    toDelete.add(item);
+            }
+
+            items.removeAll(toDelete);
+        } else {
+            log.error("IN deleteFromCart - This username can't delete items from this cart");
+            throw new NotAccessException("Username is invalid");
         }
-        items.removeAll(toDelete);
+
         log.info("IN deleteFromCart - cart with id: {} successfully delete {} items", cartId, itemIds.size());
         return cartRepository.save(dbCart);
     }
@@ -122,10 +159,12 @@ public class CartServiceImpl implements CartService {
         cart.setCartId(UUID.randomUUID());
         Collection<Item> items = cart.getItems();
         Collection<Item> cartsItem = new LinkedList<>();
+
         for (Item item : items) {
             Item savedItem = itemRepository.findItemByItemId(item.getItemId());
             cartsItem.add(savedItem);
         }
+
         cart.setItems(cartsItem);
         log.info("IN saveCart - new cart with id: {} successfully added", cart.getCartId());
         Cart savedCart = cartRepository.insert(cart);
@@ -140,27 +179,47 @@ public class CartServiceImpl implements CartService {
      * @return Updated cart.
      */
     @Override
-    public Cart updateCart(UUID cartId, Cart cart) throws IllegalStateException, EntityNotFoundException, ValidationException {
+    public Cart updateCart(UUID cartId, Cart cart) throws NotValidUsernameException, EntityNotFoundException, NotAccessException {
+
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String username;
+
+        if (principal instanceof UserDetails) {
+            username = ((UserDetails) principal).getUsername();
+        } else {
+            username = principal.toString();
+        }
+
         if (cart == null) {
             log.error("IN updateCart - cart is null");
-            throw new ValidationException("cart is null");
+            throw new NotValidUsernameException("cart is null");
         }
+
         Cart dbCart = cartRepository.findCartByCartId(cartId);
+
         if (dbCart == null) {
             log.error("IN updateCart - no cart found by id: {}", cartId);
             throw new EntityNotFoundException("cart not found");
         }
-        Collection<Item> cartsItem = new LinkedList<>();
-        if (cart.getItems() != null) {
-            for (Item item : cart.getItems()) {
-                Item savedItem = itemRepository.findItemByItemId(item.getItemId());
-                cartsItem.add(savedItem);
-            }
-            dbCart.setItems(cartsItem);
-        }
-        dbCart.setUserId(cart.getUserId());
 
-        log.info("IN updateCart - cart with id: {} successfully edited ", cartId);
+        User user = userRepository.findUserByUserId(dbCart.getUserId());
+
+        if (Objects.equals(username, user.getMail())) {
+            Collection<Item> cartsItem = new LinkedList<>();
+
+            if (cart.getItems() != null) {
+                for (Item item : cart.getItems()) {
+                    Item savedItem = itemRepository.findItemByItemId(item.getItemId());
+                    cartsItem.add(savedItem);
+                }
+                dbCart.setItems(cartsItem);
+            }
+
+            log.info("IN updateCart - cart with id: {} successfully edited ", cartId);
+        } else {
+            log.error("IN updateCart - This username can't update this cart");
+            throw new NotAccessException("Username is invalid");
+        }
         return cartRepository.save(dbCart);
     }
 
@@ -170,8 +229,32 @@ public class CartServiceImpl implements CartService {
      * @param cartId This is cart's id which needed to delete.
      */
     @Override
-    public void deleteCartById(UUID cartId) {
-        cartRepository.deleteByCartId(cartId);
-        log.info("IN deleteCartById - cart with id: {} successfully deleted", cartId);
+    public void deleteCartById(UUID cartId) throws EntityNotFoundException, NotAccessException {
+
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String username;
+
+        if (principal instanceof UserDetails) {
+            username = ((UserDetails) principal).getUsername();
+        } else {
+            username = principal.toString();
+        }
+
+        Cart dbCart = cartRepository.findCartByCartId(cartId);
+
+        if (dbCart == null) {
+            log.error("IN updateCart - no cart found by id: {}", cartId);
+            throw new EntityNotFoundException("cart not found");
+        }
+
+        User user = userRepository.findUserByUserId(dbCart.getUserId());
+
+        if (Objects.equals(username, user.getMail())) {
+            cartRepository.deleteByCartId(cartId);
+            log.info("IN delete - cart with id: {} successfully deleted", cartId);
+        } else {
+            log.error("IN updateCart - This username can't update this cart");
+            throw new NotAccessException("Username is invalid");
+        }
     }
 }
